added all functions in all pages

+++ favourite.js
@@ -42,6 +42,20 @@ async function fetchMealById(id) {
     }
 }
 
+//get recipe with id
+async function fetchRecipe(id) {
+    try {
+        const res = await fetch(idBaseURL + id)
+        const data = await res.json()
+        //find a way to return an array of objects
+        //console.log(data);
+        UI.createRecipe(data)
+        return data
+    } catch (err) {
+        console.log(err);
+    }
+}
+
 //LOAD FUNCTIONS ON WINDOW LOAD
 document.addEventListener('DOMContentLoaded', () => {
     //filter favourite meals by category
@@ -56,6 +70,11 @@ document.addEventListener('DOMContentLoaded', () => {
 let categories = document.querySelectorAll('.category');
 let grid = document.querySelector('.grid-favourite')
 let body = document.querySelector('body')
+let mealCategoriesSection = document.querySelector('.meal-categories-section')
+let recipeSection = document.querySelector('.recipe-section')
+let closeBtn = document.querySelector('.close-btn')
+let favouriteTitle = document.querySelector('.favourite-title')
+let container = document.querySelector('.container')
 
 
 //REPRESENTS A MEAL
@@ -71,6 +90,65 @@ class MealCard {
 
 //HANDLES UI TASKS
 class UI {
+
+    //create recipe
+    static createRecipe(data) {
+        recipeSection.innerHTML = `
+            <div class="close-btn"><i class="fas fa-times"></i></div>
+            <div class="recipe-image-container">
+                <img src="${ data.meals[0].strMealThumb }" alt="${ data.meals[0].strMeal }">
+            </div>
+            <div class="recipe-info-container">
+                <h3 class="recipe-title active">
+                ${ data.meals[0].strMeal }
+                </h3>
+                <div class="recipe-category-container">
+                    <p class="recipe-category active">Category:</p>
+                    <p class="recipe-category-result">${ data.meals[0].strCategory }</p>
+                </div>
+                <div class="recipe-area-container">
+                    <p class="recipe-area active">Area:</p>
+                    <p class="recipe-area-result">${ data.meals[0].strArea }</p>
+                </div>
+                <div class="heart-container" id="${ data.meals[0].idMeal }">
+                    <i class="fas fa-heart heart-2 pink"></i>
+                </div>
+                <div class="ingredients-container">
+                    <h4 class="ingredients-title active">Ingredients</h4>
+                    <ul class="ingredients-list">
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient1 }</span><span class="ingr-measure">${ data.meals[0].strMeasure1 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient2 }</span><span class="ingr-measure">${ data.meals[0].strMeasure2 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient3 }</span><span class="ingr-measure">${ data.meals[0].strMeasure3 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient4 }</span><span class="ingr-measure">${ data.meals[0].strMeasure4 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient5 }</span><span class="ingr-measure">${ data.meals[0].strMeasure5 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient6 }</span><span class="ingr-measure">${ data.meals[0].strMeasure6 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient7 }</span><span class="ingr-measure">${ data.meals[0].strMeasure7 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient8 }</span><span class="ingr-measure">${ data.meals[0].strMeasure8 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient9 }</span><span class="ingr-measure">${ data.meals[0].strMeasure9 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient10 }</span><span class="ingr-measure">${ data.meals[0].strMeasure10 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient11 }</span><span class="ingr-measure">${ data.meals[0].strMeasure11 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient12 }</span><span class="ingr-measure">${ data.meals[0].strMeasure12 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient13 }</span><span class="ingr-measure">${ data.meals[0].strMeasure13 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient14 }</span><span class="ingr-measure">${ data.meals[0].strMeasure14 }</span></li><li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient15 }</span><span class="ingr-measure">${ data.meals[0].strMeasure15 }</span></li><li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient16 }</span><span class="ingr-measure">${ data.meals[0].strMeasure16 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient17 }</span><span class="ingr-measure">${ data.meals[0].strMeasure17 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient18 }</span><span class="ingr-measure">${ data.meals[0].strMeasure18 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient19 }</span><span class="ingr-measure">${ data.meals[0].strMeasure19 }</span></li>
+                        <li class="ingredient"><span class="ingr-title">${ data.meals[0].strIngredient20 }</span><span class="ingr-measure">${ data.meals[0].strMeasure20 }</span></li>
+                    </ul>
+                </div>
+                <div class="instructions-container">
+                    <div class="instructions-title active">Instructions</div>
+                    <p class="instructions">
+                        "${ data.meals[0].strInstructions }"
+                    </p>
+                    <p class="note active">Buon Appetito!</p>
+                    <div class="video">
+                    <iframe class="video" id="player" type="text/html" width="640" height="390" src="https://www.youtube.com/embed/${ data.meals[0].strYoutube.slice(32) }" frameborder="0"></iframe>
+                    </div>
+                </div>
+            </div>
+            `
+    }
 
     //changes category background color on click
     static changeCategoryColor() {
@@ -234,11 +312,24 @@ class Storage {
 
 //remove meal on heart click
 body.addEventListener('click', (e) => {
-    if (e.target.classList.contains('fa-heart') && e.target.classList.contains('fas')) {
+    if (e.target.classList.contains('fa-heart') && e.target.classList.contains('fas') && e.target.parentElement.parentElement.parentElement.classList.contains('grid-item') === true) {
         e.target.parentElement.parentElement.parentElement.remove()
         //remove favourite meal to/from storage
         let id = e.target.parentElement.parentElement.parentElement.id
         Storage.removeMealFromLS(id)
+
+    }
+})
+
+//remove meal on heart click on RECIPE page
+body.addEventListener('click', (e) => {
+    if (e.target.classList.contains('fa-heart') && e.target.classList.contains('fas') && e.target.parentElement.parentElement.parentElement.classList.contains('grid-item') === false) {
+        //remove pink color
+        e.target.classList.remove('pink')
+        //remove favourite meal to/from storage
+        let id = e.target.parentElement.id
+        Storage.removeMealFromLS(id)
+        location.reload()
     }
 })
 
@@ -250,7 +341,48 @@ categories.forEach((category) => {
     })
 })
 
+//get meal id on grid card click
+grid.addEventListener('click', (e) => {
+    if (e.target.classList.contains('grid-item')) {
+        fetchRecipe(e.target.id)
+        console.log(e.target.id);
+        //remove none class from recipe section
+        recipeSection.classList.remove('none')
+        //add none class to favourite title
+        favouriteTitle.classList.add('none')
+        //add none class to favourite grid
+        grid.classList.add('none')
+        //scroll to top of recipe
+        container.scrollTo(0, 0)
+
+    } else if (e.target.parentElement.parentElement.classList.contains('grid-item')) {
+        fetchRecipe(e.target.parentElement.parentElement.id)
+        console.log(e.target.parentElement.parentElement.id);
+        //remove none class from recipe section
+        recipeSection.classList.remove('none')
+        //add none class to favourite title
+        favouriteTitle.classList.add('none')
+        //add none class to favourite grid
+        grid.classList.add('none')
+        //scroll to top of recipe
+        container.scrollTo(0, 0)
+    }
+})
+
+//close button  listener
+recipeSection.addEventListener('click', (e) => {
+    if (e.target.classList.contains('fa-times')) {
+        recipeSection.classList.add('none')
+        //remove none class to favourite title
+        favouriteTitle.classList.remove('none')
+        //remove none class to favourite grid
+        grid.classList.remove('none')
+    }
+})
+
 
 //TO DO:
 
-//1. get meal id on meal click
+//1. swap \r\ in recipe for <br> using regex?
+//2. fix the heart function on the recipe page
+//3. fix heart on the favourite recipe page (when removing recipe from favourites from recipe page, the card is still visible on favourites until you refresh page)
